+++ src/models/Board.ts
@@ -1,7 +1,7 @@
 import { KeyCode } from "../KeyCode";
 import { wordBank } from "../wordBank";
 import { getRandomItem } from "../getRandomItem";
-import { Word } from "../Word";
+import { Word } from "./Word";
 import { Letter } from "../Letter";
 import { SuccessStatus } from "../SuccessStatus";
